refactored db and added sessions
