+++ backend/package.json
@@ -18,21 +18,28 @@
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
+    "express-session": "^1.17.3",
     "http": "^0.0.1-security",
     "https": "^1.0.0",
     "jsonwebtoken": "^9.0.2",
+    "passport": "^0.7.0",
+    "passport-local": "^1.0.0",
     "pg": "^8.11.3",
     "ts-node-dev": "^2.0.0",
     "typescript": "^5.2.2",
     "ws": "^8.14.2"
   },
   "devDependencies": {
+    "@types/bcrypt": "^5.0.2",
     "@types/chai": "^4.3.9",
     "@types/cors": "^2.8.16",
     "@types/express": "^4.17.20",
+    "@types/express-session": "^1.17.10",
     "@types/jsonwebtoken": "^9.0.5",
     "@types/mocha": "^10.0.3",
     "@types/node": "^20.8.7",
+    "@types/passport": "^1.0.16",
+    "@types/passport-local": "^1.0.38",
     "@types/pg": "^8.10.9",
     "@types/sinon": "^10.0.20",
     "@types/ws": "^8.5.8",
