+++ backend/src/handlers/login-handler.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from 'express';
-import { validateLogin } from '../db/db';
+import { validateLogin, createAccount } from '../db/login-db';
 
 const loginHandler = async (req: Request, res: Response) => {
   const { username, password } = req.body;
@@ -16,7 +16,30 @@ const loginHandler = async (req: Request, res: Response) => {
   }
 };
 
+const registerHandler = async (req: Request, res: Response) => { 
+  const { username, password, emailAddress, firstName, lastName} = req.body;
+  // First sanitize emailAddress, username and password
+  try {
+    // Confirm that confirmPassword and password are the same
+    if (!await createAccount(username, password, emailAddress, firstName, lastName)) {
+      res.status(401).json({ error: 'User already exists!' });
+    } else {
+      res.status(200).json({ message: 'Account creation successful' });
+    }
+    }
+  catch (error) {
+    console.error('Error during account creation: ', error);
+    res.status(500).json({ error: 'Login failed' });
+  }
+}
+
+const userInfoHandler = async (req: Request, res: Response) => {
+  console.log(req.user); // FIXME ADAM
+  res.status(200).send();
+}
+
 export {
-  // eslint-disable-next-line import/prefer-default-export
   loginHandler,
+  registerHandler,
+  userInfoHandler,
 };
