+++ backend/src/routes/routes.ts
@@ -1,15 +1,16 @@
 import { Router } from 'express';
-import { loginHandler } from '../handlers/login-handler';
+import { loginHandler, registerHandler, userInfoHandler } from '../handlers/login-handler';
 import { ledOn } from '../handlers/led-handler';
 import { claimKeyboard, authorizeKeyboard } from '../handlers/auth-handler';
-import { registerHandler } from '../handlers/create-account-handler';
+import { authenticate, isAuthenticated } from './middleware';
 
 const router = Router();
 
-router.post('/login', loginHandler);
+router.post('/login', authenticate, loginHandler);
 router.post('/register', registerHandler);
 router.post('/led', ledOn);
 router.post('/claim', claimKeyboard);
 router.get('/authorize', authorizeKeyboard);
+router.get('/getUserInfo', isAuthenticated, userInfoHandler);
 
 export default router;
