+++ backend/src/types/express/index.d.ts
@@ -1,0 +1,7 @@
+declare namespace Express {
+  export interface User {
+    id: string;
+    username: string;
+    email: string;
+  }
+}
